Move some functions to helpers.py
